+++ acpi_intel_sst.c
@@ -1,5 +1,5 @@
 /*-
- * SPDX-License-Identifier: BSD-2-Clause
+ * SPDX-License-Identifier: BSD-3-Clause
  *
  * Copyright (c) 2026 FreeBSD Foundation
  * All rights reserved.